Fixed cake, added vscode workspace configs and tweaked symbol and documentation settings for the package
To stop cake from throwing exceptions apparently we have to not take release notes for granted and remove
implicit collection initializers from the reportgenerator task.
Specifying documentation output filepath seems to be buggy right now, so we dial it back to the default by
setting GenerateDocumentationFile to "true", which defaults the output to the output directory and the name
of the project, which is fine and probably best for our needs.
To prevent certain movements in the dotnet/sdk development from implicitly breaking our packages,
we pin down the DebugType of the package to XPlat external PDBs.